added some sort of firebase authentication
