+++ firebase.js
@@ -1,0 +1,11 @@
+const firebase = require("firebase-admin");
+const credsUnStruct = process.env.FIREBASE_CREDS;
+const creds = JSON.parse(credsUnStruct);
+
+firebase.initializeApp({
+    credential: firebase.credential.cert(creds),
+    databaseURL: 'https://fir-gptletters.firebaseio.com',
+})
+
+
+module.exports = firebase;
