+++ package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "main": "server.js",
-
   "scripts": {
     "client-install": "npm install --prefix client",
     "start": "node server.js",
@@ -12,9 +11,6 @@
     "dev": "concurrently \"npm run server\" \"npm run client\"",
     "heroku-postbuild": "NPM_CONFIG_PRODUCTION=false npm install --prefix client && npm run build --prefix client"
   },
-
-
-
   "dependencies": {
     "body-parser": "^1.20.1",
     "buffer": "^6.0.3",
@@ -24,12 +20,13 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "file-saver": "^2.0.5",
+    "firebase-admin": "^11.5.0",
     "form-data": "^4.0.0",
     "fs": "^0.0.1-security",
     "multer": "^1.4.5-lts.1",
-    "openai": "^3.1.0"
+    "openai": "^3.1.0",
+    "react-alert": "^7.0.3"
   },
-
   "devDependencies": {
     "nodemon": "^2.0.20"
   }
